lr1/requirements.txt
PySide6>=6.0.0
requests>=2.25.0
python-dotenv>=0.19.0
pytest>=6.0.0
